# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk Sistem Data Induk Siswa SMK",
        "title": "API Siswa Induk",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@siswa.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/orang-tua/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update parent data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orang Tua"
                ],
                "summary": "Update parent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Parent data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateOrangTuaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.OrangTuaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete parent data",
                "tags": [
                    "Orang Tua"
                ],
                "summary": "Delete parent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/pendidikan/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update previous education record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pendidikan"
                ],
                "summary": "Update education",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Education Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Education data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdatePendidikanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.PendidikanResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete previous education record",
                "tags": [
                    "Pendidikan"
                ],
                "summary": "Delete education",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Education Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/siswa/{id}/orang-tua": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create parent for a student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orang Tua"
                ],
                "summary": "Create parent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Parent data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateOrangTuaRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.OrangTuaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/siswa/{id}/pendidikan": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add previous education record to student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pendidikan"
                ],
                "summary": "Add previous education",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Student ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Education data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreatePendidikanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/responses.PendidikanResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.CreateOrangTuaRequest": {
            "type": "object",
            "required": [
                "nama",
                "tipe"
            ],
            "properties": {
                "alamat": {
                    "type": "string",
                    "example": "Jl. Merdeka No. 123"
                },
                "kewarganegaraan": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "Indonesia"
                },
                "masih_hidup": {
                    "type": "boolean",
                    "example": true
                },
                "nama": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "Budi Santoso"
                },
                "no_telepon": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "081234567890"
                },
                "pekerjaan": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "Wiraswasta"
                },
                "pendidikan_terakhir": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "S1"
                },
                "penghasilan_bulanan": {
                    "type": "number",
                    "example": 5000000
                },
                "tanggal_lahir": {
                    "type": "string",
                    "example": "1975-03-20"
                },
                "tempat_lahir": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "Jakarta"
                },
                "tipe": {
                    "type": "string",
                    "enum": [
                        "ayah",
                        "ibu"
                    ],
                    "example": "ayah"
                }
            }
        },
...
        "requests.CreatePendidikanRequest": {
            "type": "object",
            "required": [
                "asal_sekolah",
                "kelas_diterima",
                "tanggal_diterima",
                "tipe"
            ],
            "properties": {
                "alamat_sekolah": {
                    "type": "string",
                    "example": "Jl. Pendidikan No. 1"
                },
                "alasan_pindah": {
                    "type": "string",
                    "example": ""
                },
                "asal_sekolah": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "SMP Negeri 1 Bandung"
                },
                "kelas_diterima": {
                    "type": "string",
                    "enum": [
                        "X",
                        "XI",
                        "XII"
                    ],
                    "example": "X"
                },
                "no_ijazah": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "DN-01/MI-2024"
                },
                "no_skhun": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "DN-01/SK-2024"
                },
                "tanggal_diterima": {
                    "type": "string",
                    "example": "2024-07-15"
                },
                "tanggal_ijazah": {
                    "type": "string",
                    "example": "2024-06-10"
                },
                "tanggal_skhun": {
                    "type": "string",
                    "example": "2024-06-10"
                },
                "tipe": {
                    "type": "string",
                    "enum": [
                        "siswa_baru",
                        "pindahan"
                    ],
                    "example": "siswa_baru"
                }
            }
        },
...
        "requests.UpdateOrangTuaRequest": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string",
                    "example": "Jl. Merdeka No. 123"
                },
                "kewarganegaraan": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "Indonesia"
                },
                "masih_hidup": {
                    "type": "boolean",
                    "example": true
                },
                "nama": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "Budi Santoso"
                },
                "no_telepon": {
                    "type": "string",
                    "maxLength": 20,
                    "example": "081234567890"
                },
                "pekerjaan": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "Wiraswasta"
                },
                "pendidikan_terakhir": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "S1"
                },
                "penghasilan_bulanan": {
                    "type": "number",
                    "example": 5000000
                },
                "tanggal_lahir": {
                    "type": "string",
                    "example": "1975-03-20"
                },
                "tempat_lahir": {
                    "type": "string",
                    "maxLength": 100,
                    "example": "Jakarta"
                },
                "tipe": {
                    "type": "string",
                    "enum": [
                        "ayah",
                        "ibu"
                    ],
                    "example": "ayah"
                }
            }
        },
        "requests.UpdatePendidikanRequest": {
            "type": "object",
            "properties": {
                "alamat_sekolah": {
                    "type": "string",
                    "example": "Jl. Pendidikan No. 1"
                },
                "alasan_pindah": {
                    "type": "string",
                    "example": ""
                },
                "asal_sekolah": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "SMP Negeri 1 Bandung"
                },
                "kelas_diterima": {
                    "type": "string",
                    "enum": [
                        "X",
                        "XI",
                        "XII"
                    ],
                    "example": "X"
                },
                "no_ijazah": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "DN-01/MI-2024"
                },
                "no_skhun": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "DN-01/SK-2024"
                },
                "tanggal_diterima": {
                    "type": "string",
                    "example": "2024-07-15"
                },
                "tanggal_ijazah": {
                    "type": "string",
                    "example": "2024-06-10"
                },
                "tanggal_skhun": {
                    "type": "string",
                    "example": "2024-06-10"
                },
                "tipe": {
                    "type": "string",
                    "enum": [
                        "siswa_baru",
                        "pindahan"
                    ],
                    "example": "siswa_baru"
                }
            }
        },
...
        "responses.OrangTuaResponse": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kewarganegaraan": {
                    "type": "string"
                },
                "masih_hidup": {
                    "type": "boolean"
                },
                "nama": {
                    "type": "string"
                },
                "no_telepon": {
                    "type": "string"
                },
                "pekerjaan": {
                    "type": "string"
                },
                "pendidikan_terakhir": {
                    "type": "string"
                },
                "penghasilan_bulanan": {
                    "type": "number"
                },
                "tanggal_lahir": {
                    "type": "string"
                },
                "tempat_lahir": {
                    "type": "string"
                },
                "tipe": {
                    "type": "string"
                }
            }
        },
...
        "responses.PendidikanResponse": {
            "type": "object",
            "properties": {
                "alamat_sekolah": {
                    "type": "string"
                },
                "alasan_pindah": {
                    "type": "string"
                },
                "asal_sekolah": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kelas_diterima": {
                    "type": "string"
                },
                "no_ijazah": {
                    "type": "string"
                },
                "no_skhun": {
                    "type": "string"
                },
                "tanggal_diterima": {
                    "type": "string"
                },
                "tanggal_ijazah": {
                    "type": "string"
                },
                "tanggal_skhun": {
                    "type": "string"
                },
                "tipe": {
                    "type": "string"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\" to authenticate",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}